pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multi-time"
version = "0.2.0"
description = "Comprehensive multi-frequency time series analysis toolkit built on the sktime ecosystem"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "sktime>=0.26",
    "statsmodels>=0.14",
    "scipy>=1.11",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
forecasting = [
    "pmdarima>=2.0",
    "tbats>=1.1",
]
deep-learning = [
    "pytorch-forecasting>=1.0",
    "torch>=2.0",
    "lightning>=2.0",
]
probabilistic = [
    "skpro>=2.0",
]
visualization = [
    "matplotlib>=3.7",
]
all = [
    "multi-time[forecasting,probabilistic,visualization]",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3",
    "matplotlib>=3.7",
]

[project.urls]
Homepage = "https://github.com/docxology/multi-time"
Repository = "https://github.com/docxology/multi-time.git"

[tool.hatch.build.targets.wheel]
packages = ["src/multi_time"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
norecursedirs = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
